[SPARK-47304][SQL][TESTS] Distribute tests from `DataFrameSuite` to more specific suites
### What changes were proposed in this pull request?
In the PR, I propose to move some tests from `DataFrameSuite` out, and distribute them across more specific tests suites and new one `DataFrameShowSuite`.

### Why are the changes needed?
1. Improve maintainability of `DataFrameSuite`
2. Speed up execution of the test suite. Execution time dropped to < 1min.

### Does this PR introduce _any_ user-facing change?
No.

### How was this patch tested?
By running the modified test suites.

### Was this patch authored or co-authored using generative AI tooling?
No.

Closes #45392 from MaxGekk/split-DataFrameSuite.

Authored-by: Max Gekk <max.gekk@gmail.com>
